LayoutTests/platform/ios-sim-deprecated/fast/dom/Document/CaretRangeFromPoint/basic-expected.txt

CONSOLE MESSAGE: line 38: TypeError: null is not an object (evaluating 'range.startContainer')
xxxxx xxxxx xxxxx xxxxx xxxxx xxxxx xxxxx xxxxx xxxxx xxxxx xxxxx xxxxx xxxxx xxxxx xxxxx xxxxx
